Drop wheezy variant
Full support for wheezy ended on April 26, 2016. LTS support will drop
in May 2018:

https://wiki.debian.org/DebianReleases#Production_Releases

We may want to wait until the LTS drop date. However, given the recent
release of Stretch the timing seems appropriate to do it sooner.

Also worth noting is that LTS is not supported by the Debian Security
team but by volunteers and interested companies:

https://wiki.debian.org/LTS